refactor into editor file
